GPII-3513: Fix helm-initializer not being triggered
When there are 2 successive runs when cluster is (re-)created and then
cre-created again immediatelly, tiller_status was evaluated as empty ""
and would not re-trigger helm-initialized deployment. Changing the
"negative", tiller-not-installed, value to random number will make sure
the vaule is always different and tiller will be redeployed when needed.